requirements.txt
# requirements.txt
python-telegram-bot[ext]
firebase-admin
python-dotenv